ticket_backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.2
Flask-Cors==4.0.0
mysql-connector-python==8.1.0
PyMySQL==1.1.0
Werkzeug==2.3.7
python-dotenv==1.0.0
flask-wtf